test: split the measure-unit tests into their own file
